fix
remove some redundant code
